+++ agentloop/__init__.py
@@ -1,4 +1,9 @@
-from agentloop.autoinstrument import InstrumentationResult, auto_instrument
+from agentloop.autoinstrument import (
+    DetectionResult,
+    InstrumentationResult,
+    auto_instrument,
+    detect_integrations,
+)
 from agentloop.client import AgentLoopClient, AgentLoopClientError
 from agentloop.decorators import trace_model, trace_tool, traceable
 from agentloop.doctor import run_doctor
@@ -6,7 +11,14 @@ from agentloop.findings import build_diagnosis
 from agentloop.otel import trace_from_otel, trace_to_otel
 from agentloop.patches import build_patch_plan
 from agentloop.replay import ReplayGates, build_replay_report
-from agentloop.runtime import get_last_error, get_runtime_config, init, reset_runtime
+from agentloop.runtime import (
+    CLEAR,
+    FinalizationError,
+    get_last_error,
+    get_runtime_config,
+    init,
+    reset_runtime,
+)
 from agentloop.tracer import (
     AgentTrace,
     current_trace,
@@ -24,9 +36,13 @@ __all__ = [
     "AgentLoopClient",
     "AgentLoopClientError",
     "AgentTrace",
+    "CLEAR",
+    "FinalizationError",
+    "DetectionResult",
     "InstrumentationResult",
     "ReplayGates",
     "auto_instrument",
+    "detect_integrations",
     "build_value_report",
     "build_diagnosis",
     "build_patch_plan",